requirements.txt
# FastAPI后端依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv>=1.0.0

# LLM客户端依赖
openai>=1.99.0
anthropic>=0.34.0
requests>=2.31.0
aiohttp>=3.9.0

# 可视化相关（已有）
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0

# 异步任务处理（可选，后续可添加）
# celery==5.3.4
# redis==5.0.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
